// frontend/attendance-app/src/types/attendance.ts
import type { FilterPagination } from "./general";

export type AttendanceStatus =
  | "PRESENT"
  | "LATE"
  | "INCOMPLETE"
  | "ABSENT"
  | "LEAVE";

export interface Attendance {
  id: number;
  userId: number;
  clockInAt: string;
  clockOutAt: string | null;
  date: string;
  status: AttendanceStatus;
  remark: string | null;
  createdAt: string;
}

export interface AttendanceRequest {
  remark?: string;
}

export interface FilterAttendance extends FilterPagination {
  startDate?: string;
  endDate?: string;
}